code gan hoan thien

+++ app/Http/Controllers/MemberController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Ve;
 use App\Models\NguoiDung;
+use App\Models\DoAnVaChiTietVe;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
@@ -79,8 +80,9 @@ class MemberController extends Controller
     public function lichSuDatVe()
     {
         // Lấy danh sách vé đặt của user hiện tại
+        $currentTime = \Carbon\Carbon::now('Asia/Ho_Chi_Minh')->format('H:i');
         $userId = Auth::id(); // Lấy ID user hiện tại
-        $lichSuDatVe = Ve::query()
+        $lichSuDatVe = Ve::query()->withTrashed()
         ->with([
             'suatChieu'=>function($qr){
                 $qr->select('id', 'phong_chieu_id', 'phim_id','ngay', 
@@ -91,11 +93,48 @@ class MemberController extends Controller
         ])
         ->where('nguoi_dung_id', $userId)->get();
         $vedahuy = Ve::onlyTrashed()->get();
-       // dd($lichSuDatVe->toArray());
-        return view('user.lichsuvedat', compact('lichSuDatVe'));
+        // $gioBatDau = \Carbon\Carbon::createFromFormat('H:i', $lichSuDatVe[2]->suatChieu->gio_bat_dau);
+        // $gioBatDauTru15Phut = $gioBatDau->subMinutes(10)->format('H:i');
+        // if($gioBatDauTru15Phut > $currentTime){
+        //     dd($lichSuDatVe->toArray(),$currentTime,$gioBatDauTru15Phut);
+        // }else{
+        //     dd('ma');
+        // }
+        return view('user.lichsuvedat', compact('lichSuDatVe','currentTime'));
+    }
+    public function chitietvedat($id){
+        $ve = Ve::query()->withTrashed()->with([
+            'chiTietVe',
+            'suatChieu' => function ($query) {
+                $query->select('id', 'phong_chieu_id', 'phim_id', DB::raw("TIME_FORMAT(gio_bat_dau,'%H:%i') as gio_bat_dau"), DB::raw("TIME_FORMAT(gio_ket_thuc,'%H:%i') as gio_ket_thuc"))
+                    ->with([
+                        'phim',
+                        'rap',
+                        'phongChieu' => function ($qr) {
+                            $qr->with('rap');
+                        }
+                    ]);
+            },
+            'maGiamGia'
+        ])
+            ->where('id', $id)->first();
+        if (empty($ve)) {
+            return abort(404, 'Trang không tồn tại!');
+        }
+        //dd($ve->toArray());
+        $food = DoAnVaChiTietVe::query()->with([
+            'food'
+        ])->where('ve_id', $ve->id)->get();
+        //dd($food->toArray());
+        $ghe = $ve->chiTietVe->pluck('seat')->groupBy('the_loai');
+        return view('user.chitietvedat', compact(['ve', 'ghe', 'food']));
     }
     public function huyVe(Request $request, string $id)
     {
+        $ve = Ve::withTrashed()->find($id);
+        if($ve->deleted_at !== null){
+            return back()->with('vedahuy','Vé đã được huỷ từ trước đó!');
+        }
 
         $curDate = \Carbon\Carbon::now('Asia/Ho_Chi_Minh')->format('Y-m-d');
         $currentTime = \Carbon\Carbon::now('Asia/Ho_Chi_Minh')->format('H:i');
@@ -119,6 +158,9 @@ class MemberController extends Controller
             $this->congxukhihuyve($ve);
         } elseif ($ve->ngay_ve_mo == $curDate) {
             if ($timegiobatdau->diffInMinutes($currentTime) > 15) {
+                if($ve->trang_thai == 0){
+                  return back()->with('cannotcancelled','Không thể huỷ vé này vì đã quét QR xác nhận vào xem phim!');
+                }
                 $ve->delete();
                 $this->congxukhihuyve($ve);
             } else {
